# openapi.yaml
openapi: 3.0.0
info:
  title: 'GeoService API'
  version: 2.2.0
servers:
  -
    url: 'http://127.0.0.1:8085'
    description: 'Local environment'
paths:
...
  /api/v1/currencies:
    get:
      tags:
        - Currencies
      operationId: get_app_currency_api_v1_currency_list
      parameters:
        -
          name: limit
          in: query
          description: 'Default 500'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: offset
          in: query
          description: 'Default 0'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: name
          in: query
          description: 'Currency name'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: code
          in: query
          description: 'Currency code'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: symbol
          in: query
          description: 'Currency symbol'
          required: false
          content:
            application/json:
              schema:
                type: string
      responses:
        '200':
          description: 'List of currencies.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CurrencyResponse'
    post:
      tags:
        - Currencies
      operationId: post_app_currency_api_v1_currency_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCurrencyActionRequest'
      responses:
        '200':
          description: 'Currency created successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyResponse'
  '/api/v1/currencies/{id}':
    put:
      tags:
        - Currencies
      operationId: put_app_currency_api_v1_currency_update
      parameters:
        -
          name: id
          in: path
          description: 'Currency uuid'
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCurrencyActionRequest'
      responses:
        '200':
          description: 'Currency updated successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyResponse'
    delete:
      tags:
        - Currencies
      operationId: delete_app_currency_api_v1_currency_delete
      parameters:
        -
          name: id
          in: path
          description: 'Currency uuid'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Currency deleted successfully.'
  /api/v1/countries:
    get:
      tags:
        - Countries
      operationId: get_app_country_api_v1_country_list
      parameters:
        -
          name: limit
          in: query
          description: 'Default 500'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: offset
          in: query
          description: 'Default 0'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: title
          in: query
          description: 'Country title'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: id
          in: query
          description: 'Country uuid'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: iso2
          in: query
          description: 'Country ISO2 code'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: iso3
          in: query
          description: 'Country ISO3 code'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: nativeTitle
          in: query
          description: 'Country native title'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: numericCode
          in: query
          description: 'Country numeric code'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: tld
          in: query
          description: 'Country top level domain'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: subRegion
          in: query
          description: 'SubRegion title'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: currencyCode
          in: query
          description: 'Currency code'
          required: false
          content:
            application/json:
              schema:
                type: string
      responses:
        '200':
          description: 'List of countries.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryResponse'
    post:
      tags:
        - Countries
      operationId: post_app_country_api_v1_country_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCountryActionRequest'
      responses:
        '200':
          description: 'Country created successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryResponse'
  '/api/v1/countries/{id}':
    put:
      tags:
        - Countries
      operationId: put_app_country_api_v1_country_update
      parameters:
        -
          name: id
          in: path
          description: 'Country uuid'
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCountryActionRequest'
      responses:
        '200':
          description: 'Country updated successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryResponse'
    delete:
      tags:
        - Countries
      operationId: delete_app_country_api_v1_country_delete
      parameters:
        -
          name: id
          in: path
          description: 'Country uuid'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Country deleted successfully.'
  /api/v1/countries/phone-codes:
    get:
      tags:
        - Countries
      operationId: get_app_country_api_v1_country_phone_codes
      parameters:
        -
          name: limit
          in: query
          description: 'Default 500'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: offset
          in: query
          description: 'Default 0'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: title
          in: query
          description: 'Country title'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: phoneCode
          in: query
          description: 'Phone code'
          required: false
          content:
            application/json:
              schema:
                type: string
      responses:
        '200':
          description: 'List of phone codes.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryResponse'
...
components:
  schemas:
...
    RegionResponse:
      required:
        - title
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
      type: object
...
    SubRegionResponse:
      required:
        - title
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        region:
          nullable: true
          oneOf:
            -
              $ref: '#/components/schemas/RegionResponse'
      type: object
...
    TimezoneResponse:
      required:
        - title
        - code
        - utc
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        code:
          type: string
        utc:
          type: string
      type: object
...
    CurrencyResponse:
      required:
        - name
        - code
        - symbol
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        symbol:
          type: string
      type: object
    CreateCurrencyActionRequest:
      required:
        - name
        - code
        - symbol
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
        code:
          type: string
          maxLength: 3
          minLength: 1
        symbol:
          type: string
          maxLength: 10
          minLength: 1
      type: object
    UpdateCurrencyActionRequest:
      required:
        - name
        - code
        - symbol
      properties:
        name:
          type: string
        code:
          type: string
        symbol:
          type: string
      type: object
    CountryResponse:
      required:
        - title
        - iso2
        - iso3
        - numericCode
        - phoneCode
        - flag
        - tld
        - longitude
        - latitude
        - timezones
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        nativeTitle:
          type: string
          nullable: true
        iso2:
          type: string
        iso3:
          type: string
        numericCode:
          type: string
        phoneCode:
          type: string
        flag:
          type: string
        tld:
          type: string
        longitude:
          type: number
          format: float
        latitude:
          type: number
          format: float
        altitude:
          type: integer
          nullable: true
        currency:
          nullable: true
          oneOf:
            -
              $ref: '#/components/schemas/CurrencyResponse'
        subRegion:
          nullable: true
          oneOf:
            -
              $ref: '#/components/schemas/SubRegionResponse'
        timezones:
          type: array
          items:
            $ref: '#/components/schemas/TimezoneResponse'
      type: object
    CreateCountryActionRequest:
      required:
        - iso2
        - iso3
        - numericCode
        - phoneCode
        - title
        - subRegion
        - currencyCode
        - flag
        - tld
        - latitude
        - longitude
        - timezones
      properties:
        iso2:
          type: string
          maxLength: 2
          minLength: 2
        iso3:
          type: string
          maxLength: 3
          minLength: 3
        numericCode:
          type: string
          maxLength: 3
          minLength: 1
        phoneCode:
          type: string
          maxLength: 20
          minLength: 1
        title:
          type: string
          maxLength: 150
          minLength: 1
        subRegion:
          type: string
          maxLength: 150
          minLength: 1
        currencyCode:
          type: string
          maxLength: 3
          minLength: 1
        flag:
          type: string
          maxLength: 100
          minLength: 1
        tld:
          type: string
          maxLength: 20
          minLength: 1
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        timezones:
          title: 'Array of timezone codes: "Europe/Oslo"'
          type: array
          items: {  }
          minItems: 1
        altitude:
          type: integer
          nullable: true
        nativeTitle:
          type: string
          maxLength: 150
          minLength: 1
          nullable: true
      type: object
    UpdateCountryActionRequest:
      properties:
        title:
          type: string
          maxLength: 150
          minLength: 1
          nullable: true
        nativeTitle:
          type: string
          maxLength: 150
          minLength: 1
          nullable: true
        phoneCode:
          type: string
          maxLength: 20
          minLength: 1
          nullable: true
        subRegion:
          type: string
          maxLength: 150
          minLength: 1
          nullable: true
        currencyCode:
          type: string
          maxLength: 3
          minLength: 1
          nullable: true
        flag:
          type: string
          maxLength: 100
          minLength: 1
          nullable: true
        tld:
          type: string
          maxLength: 20
          minLength: 1
          nullable: true
        timezones:
          type: array
          items: {  }
          minItems: 1
          nullable: true
        latitude:
          type: number
          format: float
          nullable: true
        longitude:
          type: number
          format: float
          nullable: true
        altitude:
          type: integer
          nullable: true
      type: object
...
  securitySchemes:
    Bearer:
      type: http
      bearerFormat: JWT
      scheme: bearer
security:
  -
    Bearer: []